tests/wpt/web-platform-tests/offscreen-canvas/pixel-manipulation/2d.imageData.create1.basic.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.imageData.create1.basic</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>2d.imageData.create1.basic</h1>
<p class="desc">createImageData(imgdata) exists and returns something</p>


<script>
var t = async_test("createImageData(imgdata) exists and returns something");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

_assertDifferent(ctx.createImageData(ctx.createImageData(1, 1)), null, "ctx.createImageData(ctx.createImageData(1, 1))", "null");

t.done();

});
</script>
